CSV upload working mostly now
