+++ backend/.gitignore
@@ -7,7 +7,8 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
-uploads
+tmp
+.env
 node_modules
 dist
 dist-ssr
