+++ backend/dotenv
@@ -1,0 +1,4 @@
+DB_HOST = localhost
+DB_USER = root
+DB_PASSWORD = **YOUR PASSWORD HERE!!!**
+DB_NAME = dio_membership_test
