+++ backend/package.json
@@ -1,8 +1,11 @@
 {
   "dependencies": {
     "cors": "^2.8.5",
+    "csv-parser": "^3.1.0",
+    "dotenv": "^16.4.7",
     "express": "^4.21.2",
-    "multer": "^1.4.5-lts.1"
+    "multer": "^1.4.5-lts.1",
+    "mysql2": "^3.12.0"
   },
   "scripts": {
     "start": "node index.js"
