+++ frontend/src/components/FileUploadButton.jsx
@@ -24,12 +24,14 @@ export default function InputFileUpload() {
     const selectedFile = event.target.files[0];
     setFile(selectedFile);
     handleUpload(selectedFile);
+    // IF WANT TO UPLOAD THE SAME FILE IN SAME DIRECTORY TWICE, UNCOMMENT
+    // event.target.value = '';
   };
 
   const handleUpload = (file) => {
     if (file) {
       const formData = new FormData();
-      formData.append('file', file);
+      formData.append('csv_file', file);
       console.log("going to upload file");
       fetch('http://localhost:3001/upload', {
         method: 'POST',
@@ -59,7 +61,7 @@ export default function InputFileUpload() {
       Import Attendance Data
       <VisuallyHiddenInput
         type="file"
-        // TODO: do something with the file :D :D :D 
+        accept=".csv"
         onChange={handleFileChange}
       />
       </Button>
